Creates separate Input components for Checkbox, TextField and TextAreaField
